services: audioflinger: Add 8_24_BIT and 24_BIT_PACKED supports
WARNING: This will just add the availability for 8_24_BIT and 24_BIT_PACKED format, it won't u$

At first the FastMixer Thread tries to downsample to 16_BIT for	"faster	mixing".
With directly pointing to the mFormat of the Input file	we prevent that.

On Audio Mix we	have disabled the Resampling factors for 8_24_BIT and 24_BIT_PACKED

Change-Id: I9bbbd80e18e02d2b112caf9a17248b18da3bccb7
Signed-off-by: Tristan Marsell <tristan.marsell@t-online.de>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>